Enpass icon proposal - Try 3
I removed the white border, made the circle 30x30 @ 9x,9y and aligned the keyhole to the grid with some small adjustments. This version has the keyhole cut from the circle but I have another where the keyhole is over the white circle (with no shadow) in case that is prefered.